+++ lib/dial/railtie.rb
@@ -4,6 +4,7 @@ require "rails"
 require "active_record"
 require "prosopite"
 
+require_relative "prosopite"
 require_relative "middleware"
 require_relative "prosopite_logger"