+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 
+- Fix parsing of multiline N+1 query logs
 - Truncate summary N+1 queries to 100 chars
 
 ## [0.3.0] - 2025-05-02